Organize common rewards Add some common values Add cosine similarity Typo change in PlayerData string Improve replay converter Remove ShootBallReward
