+++ rlgym/utils/common_values.py
@@ -1,5 +1,12 @@
 ORANGE_GOAL_CENTER = (0, 5120, 642.775 / 2)
 BLUE_GOAL_CENTER = (0, -5120, 642.775 / 2)
+
+# Often more useful than center
+ORANGE_GOAL_BACK = (0, 6000, 642.775 / 2)
+BLUE_GOAL_BACK = (0, -6000, 642.775 / 2)
+
+BALL_RADIUS = 92.75
+
 BLUE_TEAM = 0
 ORANGE_TEAM = 1
 NUM_ACTIONS = 8
