+++ rlgym/utils/reward_functions/__init__.py
@@ -1,5 +1,3 @@
 from .reward_function import RewardFunction
-from .shoot_ball_reward import ShootBallReward
 from .default_reward import DefaultReward
 from .combined_reward import CombinedReward
-from .common_rewards import *
