--- rlgym/utils/reward_functions/common_rewards.py
@@ -1,239 +1,0 @@
-from abc import abstractmethod
-
-import numpy as np
-
-from rlgym.utils import math
-from rlgym.utils.common_values import BLUE_TEAM, ORANGE_GOAL_CENTER, BLUE_GOAL_CENTER, ORANGE_TEAM
-from rlgym.utils.gamestates import GameState, PlayerData
-from rlgym.utils.reward_functions import RewardFunction
-
-
-class EventReward(RewardFunction):
-    def __init__(self, goal=0., team_goal=0., concede=-0., touch=0., shot=0., save=0., demo=0.):
-        """
-        :param goal: reward for goal scored by player.
-        :param team_goal: reward for goal scored by player's team.
-        :param concede: reward for goal scored by opponents. Should be negative if used as punishment.
-        :param touch: reward for touching the ball.
-        :param shot: reward for shooting the ball (as detected by Rocket League).
-        :param save: reward for saving the ball (as detected by Rocket League).
-        :param demo: reward for demolishing a player.
-        """
-        super().__init__()
-        self.weights = np.array([goal, team_goal, concede, touch, shot, save, demo])
-
-        # Need to keep track of last registered value to detect changes
-        self.last_registered_values = {}
-
-    @staticmethod
-    def _extract_values(player: PlayerData, state: GameState):
-        if player.team_num == BLUE_TEAM:
-            team, opponent = state.blue_score, state.orange_score
-        else:
-            team, opponent = state.orange_score, state.blue_score
-
-        return np.array([player.match_goals, team, opponent, player.ball_touched, player.match_shots,
-                         player.match_saves, player.match_demolishes])
-
-    def reset(self, initial_state: GameState, optional_data=None):
-        # Update every reset since rocket league may crash and be restarted with clean values
-        self.last_registered_values = {}
-        for player in initial_state.players:
-            self.last_registered_values[player.car_id] = self._extract_values(player, initial_state)
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray, optional_data=None):
-        old_values = self.last_registered_values[player.car_id]
-        new_values = self._extract_values(player, state)
-
-        diff_values = new_values - old_values
-        diff_values[diff_values < 0] = 0  # We only care about increasing values
-
-        reward = np.dot(self.weights, diff_values)
-
-        self.last_registered_values[player.car_id] = new_values
-        return reward
-
-
-class DistanceBallToGoalReward(RewardFunction):
-    def __init__(self, own_goal=False):
-        super().__init__()
-        self.own_goal = own_goal
-
-    def reset(self, initial_state: GameState):
-        pass
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        if player.team_num == BLUE_TEAM and not self.own_goal \
-                or player.team_num == ORANGE_TEAM and self.own_goal:
-            objective = np.array(ORANGE_GOAL_CENTER)
-        else:
-            objective = np.array(BLUE_GOAL_CENTER)
-        objective[1] *= 6000 / 5120  # Use back of net
-
-        dist = np.linalg.norm(state.ball.position - objective) - 786  # Compensate for moving objective to back of net
-        return np.exp(-0.5 * dist / 100)  # From https://arxiv.org/abs/2105.12196
-
-
-class DistancePlayerToBallReward(RewardFunction):
-    def reset(self, initial_state: GameState):
-        pass
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        dist = np.linalg.norm(player.car_data.position - state.ball.position) - 94  # Compensate for ball radius
-        return np.exp(-0.5 * dist / 100)  # From https://arxiv.org/abs/2105.12196
-
-
-class VelocityPlayerToBallReward(RewardFunction):
-    def __init__(self, use_scalar_projection=True):
-        super().__init__()
-        self.use_scalar_projection = use_scalar_projection
-
-    def reset(self, initial_state: GameState):
-        pass
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        vel = player.car_data.linear_velocity
-        pos_diff = state.ball.position - player.car_data.position
-        if self.use_scalar_projection:
-            # Vector version of v=d/t <=> t=d/v <=> 1/t=v/d
-            # Max value should be max_speed / ball_radius = 2300 / 94 = 24.5
-            # Used to guide the agent towards the ball
-            inv_t = math.scalar_projection(vel, pos_diff)
-            return inv_t
-        else:
-            # Regular component velocity
-            norm_pos_diff = pos_diff / np.linalg.norm(pos_diff)
-            vel /= 100  # uu = cm -> m
-            return float(np.dot(norm_pos_diff, vel))
-
-
-class VelocityBallToGoalReward(RewardFunction):
-    def __init__(self, own_goal=False, use_scalar_projection=True):
-        super().__init__()
-        self.own_goal = own_goal
-        self.use_scalar_projection = use_scalar_projection
-
-    def reset(self, initial_state: GameState):
-        pass
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        if player.team_num == BLUE_TEAM and not self.own_goal \
-                or player.team_num == ORANGE_TEAM and self.own_goal:
-            objective = np.array(ORANGE_GOAL_CENTER)
-        else:
-            objective = np.array(BLUE_GOAL_CENTER)
-        objective[1] *= 6000 / 5120  # Use back of net instead to prevent exploding reward
-
-        vel = state.ball.linear_velocity
-        pos_diff = objective - state.ball.position
-        if self.use_scalar_projection:
-            # Vector version of v=d/t <=> t=d/v <=> 1/t=v/d
-            # Max value should be max_speed / ball_radius = 2300 / 94 = 24.5
-            # Used to guide the agent towards the ball
-            inv_t = math.scalar_projection(vel, pos_diff)
-            return inv_t
-        else:
-            # Regular component velocity
-            norm_pos_diff = pos_diff / np.linalg.norm(pos_diff)
-            vel /= 100  # uu/s = cm/s -> m/s
-            return float(np.dot(norm_pos_diff, vel))
-
-
-class VelocityReward(RewardFunction):
-    # Simple reward function to ensure the model is training.
-    def __init__(self, negative=False):
-        super().__init__()
-        self.negative = negative
-
-    def reset(self, initial_state: GameState):
-        pass
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        return np.linalg.norm(player.car_data.linear_velocity) / 100 * (1 - 2 * self.negative)
-
-
-class SaveBoostReward(RewardFunction):
-    def reset(self, initial_state: GameState):
-        pass
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        # 1 reward for each frame with 100 boost, sqrt because 0->20 makes bigger difference than 80->100
-        return np.sqrt(player.boost_amount)
-
-
-class ConstantReward(RewardFunction):
-    def reset(self, initial_state: GameState):
-        pass
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        return 1
-
-
-class BallYCoordinateReward(RewardFunction):
-    def reset(self, initial_state: GameState):
-        pass
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        if player.team_num == BLUE_TEAM:
-            return (state.ball.position[1] / (5120 + 94)) ** 3
-        else:
-            return (state.inverted_ball.position[1] / (5120 + 94)) ** 3
-
-
-class FaceBallReward(RewardFunction):
-    def reset(self, initial_state: GameState):
-        pass
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        pos_diff = state.ball.position - player.car_data.position
-        norm_pos_diff = pos_diff / np.linalg.norm(pos_diff)
-        return float(np.dot(player.car_data.forward(), norm_pos_diff))
-
-
-class ConditionalRewardFunction(RewardFunction):
-    def __init__(self, reward_func: RewardFunction):
-        super().__init__()
-        self.reward_func = reward_func
-
-    @abstractmethod
-    def condition(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> bool:
-        raise NotImplementedError
-
-    def reset(self, initial_state: GameState):
-        pass
-
-    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        if self.condition(player, state, previous_action):
-            return self.reward_func.get_reward(player, state, previous_action)
-        return 0
-
-    def get_final_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
-        if self.condition(player, state, previous_action):
-            return self.reward_func.get_final_reward(player, state, previous_action)
-        return 0
-
-
-class RewardIfClosestToBall(ConditionalRewardFunction):
-    def __init__(self, reward_func: RewardFunction, team_only=True):
-        super().__init__(reward_func)
-        self.team_only = team_only
-
-    def condition(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> bool:
-        dist = np.linalg.norm(player.car_data.position - state.ball.position)
-        for player2 in state.players:
-            if not self.team_only or player2.team_num == player.team_num:
-                dist2 = np.linalg.norm(player2.car_data.position - state.ball.position)
-                if dist2 < dist:
-                    return False
-        return True
-
-
-class RewardIfTouchedLast(ConditionalRewardFunction):
-    def condition(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> bool:
-        return state.last_touch == player.car_id
-
-
-class RewardIfBehindBall(ConditionalRewardFunction):
-    def condition(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> bool:
-        return player.team_num == BLUE_TEAM and player.car_data.position[1] < state.ball.position[1] \
-               or player.team_num == ORANGE_TEAM and player.car_data.position[1] > state.ball.position[1]
