+++ rlgym/utils/reward_functions/common_rewards/__init__.py
@@ -1,0 +1,4 @@
+from .ball_goal_rewards import *
+from .conditional_rewards import *
+from .misc_rewards import *
+from .player_ball_rewards import *
