+++ rlgym/utils/reward_functions/common_rewards/misc_rewards.py
@@ -1,0 +1,108 @@
+import numpy as np
+
+from rlgym.utils import math
+from rlgym.utils.common_values import BLUE_TEAM, BLUE_GOAL_BACK, ORANGE_GOAL_BACK, ORANGE_TEAM
+from rlgym.utils.gamestates import GameState, PlayerData
+from rlgym.utils.reward_functions import RewardFunction
+
+
+class EventReward(RewardFunction):
+    def __init__(self, goal=0., team_goal=0., concede=-0., touch=0., shot=0., save=0., demo=0.):
+        """
+        :param goal: reward for goal scored by player.
+        :param team_goal: reward for goal scored by player's team.
+        :param concede: reward for goal scored by opponents. Should be negative if used as punishment.
+        :param touch: reward for touching the ball.
+        :param shot: reward for shooting the ball (as detected by Rocket League).
+        :param save: reward for saving the ball (as detected by Rocket League).
+        :param demo: reward for demolishing a player.
+        """
+        super().__init__()
+        self.weights = np.array([goal, team_goal, concede, touch, shot, save, demo])
+
+        # Need to keep track of last registered value to detect changes
+        self.last_registered_values = {}
+
+    @staticmethod
+    def _extract_values(player: PlayerData, state: GameState):
+        if player.team_num == BLUE_TEAM:
+            team, opponent = state.blue_score, state.orange_score
+        else:
+            team, opponent = state.orange_score, state.blue_score
+
+        return np.array([player.match_goals, team, opponent, player.ball_touched, player.match_shots,
+                         player.match_saves, player.match_demolishes])
+
+    def reset(self, initial_state: GameState, optional_data=None):
+        # Update every reset since rocket league may crash and be restarted with clean values
+        self.last_registered_values = {}
+        for player in initial_state.players:
+            self.last_registered_values[player.car_id] = self._extract_values(player, initial_state)
+
+    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray, optional_data=None):
+        old_values = self.last_registered_values[player.car_id]
+        new_values = self._extract_values(player, state)
+
+        diff_values = new_values - old_values
+        diff_values[diff_values < 0] = 0  # We only care about increasing values
+
+        reward = np.dot(self.weights, diff_values)
+
+        self.last_registered_values[player.car_id] = new_values
+        return reward
+
+
+class VelocityReward(RewardFunction):
+    # Simple reward function to ensure the model is training.
+    def __init__(self, negative=False):
+        super().__init__()
+        self.negative = negative
+
+    def reset(self, initial_state: GameState):
+        pass
+
+    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
+        return np.linalg.norm(player.car_data.linear_velocity) / 100 * (1 - 2 * self.negative)
+
+
+class SaveBoostReward(RewardFunction):
+    def reset(self, initial_state: GameState):
+        pass
+
+    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
+        # 1 reward for each frame with 100 boost, sqrt because 0->20 makes bigger difference than 80->100
+        return np.sqrt(player.boost_amount)
+
+
+class ConstantReward(RewardFunction):
+    def reset(self, initial_state: GameState):
+        pass
+
+    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
+        return 1
+
+
+class AlignBallGoal(RewardFunction):
+    def __init__(self, defense=1., offense=1.):
+        super().__init__()
+        self.defense = defense
+        self.offense = offense
+
+    def reset(self, initial_state: GameState):
+        pass
+
+    def get_reward(self, player: PlayerData, state: GameState, previous_action: np.ndarray) -> float:
+        ball = state.ball.position
+        pos = player.car_data.position
+        protecc = np.array(BLUE_GOAL_BACK)
+        attacc = np.array(ORANGE_GOAL_BACK)
+        if player.team_num == ORANGE_TEAM:
+            protecc, attacc = attacc, protecc
+
+        # Align player->ball and net->player vectors
+        defensive_reward = self.defense * math.cosine_similarity(ball - pos, pos - protecc)
+
+        # Align player->ball and player->net vectors
+        offensive_reward = self.offense * math.cosine_similarity(ball - pos, attacc - pos)
+
+        return defensive_reward + offensive_reward
