+++ rlgym/utils/reward_functions/reward_function.py
@@ -8,10 +8,6 @@ import numpy as np
 
 
 class RewardFunction(ABC):
-
-    def __init__(self):
-        pass
-
     @abstractmethod
     def reset(self, initial_state: GameState):
         """
